Fix spuriously-included test case in compile.bat
This was clumsily added in 1c4318331663b152b0b298bd2c9e5c971506a86b as a
test case for code I was writing at some point but has not been included
in the repository yet. This means the 0.6 tag is broken for anyone
trying to build from source unless this change is cherry-picked on top.
Lesson learned: keep the working tree clean!