Fixing broken behavior in the L4D2 addon system
Rather profound for the custom campaign economy & newest/TLS enjoyers.
See code comments for what this achieves, there is no sense in me trying
to write a commit description when mike is just going to rewrite it
anyway.

Thanks to bill for helping me with the RE & assistance on writing the
hooks/code and so on. I tracked down a lot of this myself, but the end
result wouldn't have been possible without his help.