Turn confixes into a feature
cvar/cmd flags that are altered using the chflags and chcmdflags
functions are now restored to their original state on unload. The exact
implementation is still up for debate, but the general idea is there,
and it works. I also added a function for unhiding commands, since
that didn't exist (for all of this time, we were trying to unhide
demo_gototick as if it were a cvar...lol)