Fix the Swamp godmode/no idle/no votes/etc bug
Turns out this was fixed in version 2112 (October 2012), so it was
fairly easy to isolate in Ghidra (I at least previously thought this
was fixed by TLS). This CDirector member represents the
finale escape state, and is used in CDirector::IsFinaleWon (and maybe
another function, don't remember anymore). The member was never cleared
after a campaign was finished, so when the Swamp (or Crash Course)
'minifinale' events ran, the game thought you were on a finale that you
just 'won', hence blocking votes/godmode/idle. 2112 fixed this by setting
the member back to 0 in CDirector::Reset, so we just set it to 0 when
quickreset is used.

Now co-op hosts won't need to restart their game after finishing
a campaign anymore!